refactor(node)!: Invert the inheritance between `Yarn` and `Npm`
Prepare for re-writing large parts of `Npm` to no more rely on the file
hierarchy of `node_modules` when constructing the dependency tree.

While the same is planned for `Yarn` too, this change is just a
refactoring step which allows to do `Npm` first.
