+++ plugins/package-managers/node/src/main/kotlin/Yarn.kt
@@ -21,6 +21,7 @@ package org.ossreviewtoolkit.plugins.packagemanagers.node
 
 import java.io.File
 import java.lang.invoke.MethodHandles
+import java.util.concurrent.ConcurrentHashMap
 
 import kotlin.time.Duration.Companion.days
 
@@ -32,17 +33,40 @@ import kotlinx.serialization.json.JsonPrimitive
 import kotlinx.serialization.json.decodeToSequence
 import kotlinx.serialization.json.jsonPrimitive
 
+import org.apache.logging.log4j.kotlin.logger
 import org.apache.logging.log4j.kotlin.loggerOf
 
 import org.ossreviewtoolkit.analyzer.AbstractPackageManagerFactory
+import org.ossreviewtoolkit.analyzer.PackageManager
+import org.ossreviewtoolkit.analyzer.PackageManagerResult
+import org.ossreviewtoolkit.model.DependencyGraph
+import org.ossreviewtoolkit.model.Identifier
+import org.ossreviewtoolkit.model.Issue
+import org.ossreviewtoolkit.model.Project
+import org.ossreviewtoolkit.model.ProjectAnalyzerResult
+import org.ossreviewtoolkit.model.Severity
 import org.ossreviewtoolkit.model.config.AnalyzerConfiguration
 import org.ossreviewtoolkit.model.config.RepositoryConfiguration
+import org.ossreviewtoolkit.model.createAndLogIssue
+import org.ossreviewtoolkit.model.readTree
+import org.ossreviewtoolkit.model.utils.DependencyGraphBuilder
 import org.ossreviewtoolkit.plugins.packagemanagers.node.utils.NodePackageManager
 import org.ossreviewtoolkit.plugins.packagemanagers.node.utils.NpmDetection
+import org.ossreviewtoolkit.plugins.packagemanagers.node.utils.NpmModuleInfo
+import org.ossreviewtoolkit.plugins.packagemanagers.node.utils.YarnDependencyHandler
+import org.ossreviewtoolkit.plugins.packagemanagers.node.utils.splitNpmNamespaceAndName
+import org.ossreviewtoolkit.utils.common.CommandLineTool
 import org.ossreviewtoolkit.utils.common.DiskCache
 import org.ossreviewtoolkit.utils.common.Os
+import org.ossreviewtoolkit.utils.common.ProcessCapture
 import org.ossreviewtoolkit.utils.common.alsoIfNull
+import org.ossreviewtoolkit.utils.common.collectMessages
+import org.ossreviewtoolkit.utils.common.fieldNamesOrEmpty
+import org.ossreviewtoolkit.utils.common.isSymbolicLink
 import org.ossreviewtoolkit.utils.common.mebibytes
+import org.ossreviewtoolkit.utils.common.realFile
+import org.ossreviewtoolkit.utils.common.stashDirectories
+import org.ossreviewtoolkit.utils.common.textValueOrEmpty
 import org.ossreviewtoolkit.utils.ort.ortDataDirectory
 
 import org.semver4j.RangesList
@@ -54,15 +78,24 @@ private val yarnInfoCache = DiskCache(
     maxCacheEntryAgeInSeconds = 7.days.inWholeSeconds
 )
 
+/** Name of the scope with the regular dependencies. */
+private const val DEPENDENCIES_SCOPE = "dependencies"
+
+/** Name of the scope with optional dependencies. */
+private const val OPTIONAL_DEPENDENCIES_SCOPE = "optionalDependencies"
+
+/** Name of the scope with development dependencies. */
+private const val DEV_DEPENDENCIES_SCOPE = "devDependencies"
+
 /**
  * The [Yarn](https://classic.yarnpkg.com/) package manager for JavaScript.
  */
-class Yarn(
+open class Yarn(
     name: String,
     analysisRoot: File,
     analyzerConfig: AnalyzerConfiguration,
     repoConfig: RepositoryConfiguration
-) : Npm(name, analysisRoot, analyzerConfig, repoConfig) {
+) : PackageManager(name, analysisRoot, analyzerConfig, repoConfig), CommandLineTool {
     class Factory : AbstractPackageManagerFactory<Yarn>("Yarn") {
         override val globsForDefinitionFiles = listOf("package.json")
 
@@ -73,7 +106,32 @@ class Yarn(
         ) = Yarn(type, analysisRoot, analyzerConfig, repoConfig)
     }
 
-    override fun hasLockfile(projectDir: File) = NodePackageManager.YARN.hasLockfile(projectDir)
+    /** Cache for submodules identified by its moduleDir absolutePath */
+    private val submodulesCache = ConcurrentHashMap<String, Set<File>>()
+
+    private val rawModuleInfoCache = mutableMapOf<Pair<File, Set<String>>, RawModuleInfo>()
+
+    private val graphBuilder by lazy { DependencyGraphBuilder(YarnDependencyHandler(this)) }
+
+    protected open fun hasLockfile(projectDir: File) = NodePackageManager.YARN.hasLockfile(projectDir)
+
+    /**
+     * Load the submodule directories of the project defined in [moduleDir].
+     */
+    protected open fun loadWorkspaceSubmodules(moduleDir: File): Set<File> {
+        val nodeModulesDir = moduleDir.resolve("node_modules")
+        if (!nodeModulesDir.isDirectory) return emptySet()
+
+        val searchDirs = nodeModulesDir.walk().maxDepth(1).filter {
+            (it.isDirectory && it.name.startsWith("@")) || it == nodeModulesDir
+        }
+
+        return searchDirs.flatMapTo(mutableSetOf()) { dir ->
+            dir.walk().maxDepth(1).filter {
+                it.isDirectory && it.isSymbolicLink() && it != dir
+            }
+        }
+    }
 
     override fun command(workingDir: File?) = if (Os.isWindows) "yarn.cmd" else "yarn"
 
@@ -87,10 +145,253 @@ class Yarn(
         // fixed minor version to be sure to get consistent results.
         checkVersion()
 
-    override fun runInstall(workingDir: File) =
+    override fun resolveDependencies(definitionFile: File, labels: Map<String, String>): List<ProjectAnalyzerResult> {
+        val workingDir = definitionFile.parentFile
+
+        return try {
+            stashDirectories(workingDir.resolve("node_modules")).use {
+                resolveDependenciesInternal(definitionFile)
+            }
+        } finally {
+            rawModuleInfoCache.clear()
+        }
+    }
+
+    /**
+     * An internally used data class with information about a module retrieved from the module's package.json. This
+     * information is further processed and eventually converted to an [NpmModuleInfo] object containing everything
+     * required by the Npm package manager.
+     */
+    private data class RawModuleInfo(
+        val name: String,
+        val version: String,
+        val dependencyNames: Set<String>,
+        val packageJson: File
+    )
+
+    // TODO: Add support for bundledDependencies.
+    private fun resolveDependenciesInternal(definitionFile: File): List<ProjectAnalyzerResult> {
+        val workingDir = definitionFile.parentFile
+
+        // Actually installing the dependencies is the easiest way to get the metadata of all transitive
+        // dependencies (i.e. their respective "package.json" files). As NPM uses a global cache, the same
+        // dependency is only ever downloaded once.
+        val installIssues = installDependencies(workingDir)
+
+        if (installIssues.any { it.severity == Severity.ERROR }) {
+            val project = runCatching {
+                parseProject(definitionFile, analysisRoot, managerName)
+            }.getOrElse {
+                logger.error { "Failed to parse project information: ${it.collectMessages()}" }
+                Project.EMPTY
+            }
+
+            return listOf(ProjectAnalyzerResult(project, emptySet(), installIssues))
+        }
+
+        val projectDirs = findWorkspaceSubmodules(workingDir).toSet() + definitionFile.parentFile
+
+        return projectDirs.map { projectDir ->
+            val issues = mutableListOf<Issue>().apply {
+                if (projectDir == workingDir) addAll(installIssues)
+            }
+
+            val project = runCatching {
+                parseProject(projectDir.resolve("package.json"), analysisRoot, managerName)
+            }.getOrElse {
+                issues += createAndLogIssue(
+                    source = managerName,
+                    message = "Failed to parse project information: ${it.collectMessages()}"
+                )
+
+                Project.EMPTY
+            }
+
+            val scopeNames = setOfNotNull(
+                // Optional dependencies are just like regular dependencies except that NPM ignores failures when
+                // installing them (see https://docs.npmjs.com/files/package.json#optionaldependencies), i.e. they are
+                // not a separate scope in ORT semantics.
+                buildDependencyGraphForScopes(
+                    project,
+                    projectDir,
+                    setOf(DEPENDENCIES_SCOPE, OPTIONAL_DEPENDENCIES_SCOPE),
+                    DEPENDENCIES_SCOPE,
+                    projectDirs,
+                    workspaceDir = workingDir
+                ),
+
+                buildDependencyGraphForScopes(
+                    project,
+                    projectDir,
+                    setOf(DEV_DEPENDENCIES_SCOPE),
+                    DEV_DEPENDENCIES_SCOPE,
+                    projectDirs,
+                    workspaceDir = workingDir
+                )
+            )
+
+            ProjectAnalyzerResult(
+                project = project.copy(scopeNames = scopeNames),
+                // Packages are set later by createPackageManagerResult().
+                packages = emptySet(),
+                issues = issues
+            )
+        }
+    }
+
+    private fun getModuleInfo(
+        moduleDir: File,
+        scopes: Set<String>,
+        projectDirs: Set<File>,
+        ancestorModuleDirs: List<File> = emptyList(),
+        ancestorModuleIds: List<Identifier> = emptyList()
+    ): NpmModuleInfo? {
+        val moduleInfo = parsePackageJson(moduleDir, scopes)
+        val dependencies = mutableSetOf<NpmModuleInfo>()
+        val packageType = managerName.takeIf { moduleDir.realFile() in projectDirs } ?: "NPM"
+
+        val moduleId = splitNpmNamespaceAndName(moduleInfo.name).let { (namespace, name) ->
+            Identifier(packageType, namespace, name, moduleInfo.version)
+        }
+
+        val cycleStartIndex = ancestorModuleIds.indexOf(moduleId)
+        if (cycleStartIndex >= 0) {
+            val cycle = (ancestorModuleIds.subList(cycleStartIndex, ancestorModuleIds.size) + moduleId)
+                .joinToString(" -> ")
+            logger.debug { "Not adding dependency '$moduleId' to avoid cycle: $cycle." }
+            return null
+        }
+
+        val pathToRoot = listOf(moduleDir) + ancestorModuleDirs
+        moduleInfo.dependencyNames.forEach { dependencyName ->
+            val dependencyModuleDirPath = findDependencyModuleDir(dependencyName, pathToRoot)
+
+            if (dependencyModuleDirPath.isNotEmpty()) {
+                val dependencyModuleDir = dependencyModuleDirPath.first()
+
+                getModuleInfo(
+                    moduleDir = dependencyModuleDir,
+                    scopes = setOf("dependencies", "optionalDependencies"),
+                    projectDirs,
+                    ancestorModuleDirs = dependencyModuleDirPath.subList(1, dependencyModuleDirPath.size),
+                    ancestorModuleIds = ancestorModuleIds + moduleId
+                )?.let { dependencies += it }
+
+                return@forEach
+            }
+
+            logger.debug {
+                "It seems that the '$dependencyName' module was not installed as the package file could not be found " +
+                    "anywhere in '${pathToRoot.joinToString()}'. This might be fine if the module is specific to a " +
+                    "platform other than the one ORT is running on. A typical example is the 'fsevents' module."
+            }
+        }
+
+        return NpmModuleInfo(
+            id = moduleId,
+            workingDir = moduleDir,
+            packageFile = moduleInfo.packageJson,
+            dependencies = dependencies,
+            isProject = moduleDir.realFile() in projectDirs
+        )
+    }
+
+    /**
+     * Retrieve all the dependencies of [project] from the given [scopes] and add them to the dependency graph under
+     * the given [targetScope]. Return the target scope name if dependencies are found; *null* otherwise.
+     */
+    private fun buildDependencyGraphForScopes(
+        project: Project,
+        workingDir: File,
+        scopes: Set<String>,
+        targetScope: String,
+        projectDirs: Set<File>,
+        workspaceDir: File? = null
+    ): String? {
+        if (excludes.isScopeExcluded(targetScope)) return null
+
+        val qualifiedScopeName = DependencyGraph.qualifyScope(project, targetScope)
+        val moduleInfo = checkNotNull(getModuleInfo(workingDir, scopes, projectDirs, listOfNotNull(workspaceDir)))
+
+        moduleInfo.dependencies.forEach { graphBuilder.addDependency(qualifiedScopeName, it) }
+
+        return targetScope.takeUnless { moduleInfo.dependencies.isEmpty() }
+    }
+
+    private fun parsePackageJson(moduleDir: File, scopes: Set<String>): RawModuleInfo =
+        rawModuleInfoCache.getOrPut(moduleDir to scopes) {
+            val packageJsonFile = moduleDir.resolve("package.json")
+            logger.debug { "Parsing module info from '${packageJsonFile.absolutePath}'." }
+            val json = packageJsonFile.readTree()
+
+            val name = json["name"].textValueOrEmpty()
+            if (name.isBlank()) {
+                logger.warn {
+                    "The '$packageJsonFile' does not set a name, which is only allowed for unpublished packages."
+                }
+            }
+
+            val version = json["version"].textValueOrEmpty()
+            if (version.isBlank()) {
+                logger.warn {
+                    "The '$packageJsonFile' does not set a version, which is only allowed for unpublished packages."
+                }
+            }
+
+            val dependencyNames = scopes.flatMapTo(mutableSetOf()) { scope ->
+                // Yarn ignores "//" keys in the dependencies to allow comments, therefore ignore them here as well.
+                json[scope].fieldNamesOrEmpty().asSequence().filterNot { it == "//" }
+            }
+
+            RawModuleInfo(
+                name = name,
+                version = version,
+                dependencyNames = dependencyNames,
+                packageJson = packageJsonFile
+            )
+        }
+
+    /**
+     * Find the directories which are defined as submodules of the project within [moduleDir].
+     */
+    private fun findWorkspaceSubmodules(moduleDir: File): Set<File> =
+        submodulesCache.getOrPut(moduleDir.absolutePath) {
+            loadWorkspaceSubmodules(moduleDir)
+        }
+
+    override fun createPackageManagerResult(projectResults: Map<File, List<ProjectAnalyzerResult>>) =
+        PackageManagerResult(projectResults, graphBuilder.build(), graphBuilder.packages())
+
+    /**
+     * Install dependencies using the given package manager command.
+     */
+    private fun installDependencies(workingDir: File): List<Issue> {
+        requireLockfile(workingDir) { hasLockfile(workingDir) }
+
+        // Install all NPM dependencies to enable NPM to list dependencies.
+        val process = runInstall(workingDir)
+
+        val lines = process.stderr.lines()
+        val issues = mutableListOf<Issue>()
+
+        // Generally forward issues from the NPM CLI to the ORT NPM package manager. Lower the severity of warnings to
+        // hints, as warnings usually do not prevent the ORT NPM package manager from getting the dependencies right.
+        lines.groupLines("npm WARN ", "npm warn ").mapTo(issues) {
+            Issue(source = managerName, message = it, severity = Severity.HINT)
+        }
+
+        // For errors, however, something clearly went wrong, so keep the severity here.
+        lines.groupLines("npm ERR! ", "npm error ").mapTo(issues) {
+            Issue(source = managerName, message = it, severity = Severity.ERROR)
+        }
+
+        return issues
+    }
+
+    protected open fun runInstall(workingDir: File): ProcessCapture =
         run(workingDir, "install", "--ignore-scripts", "--ignore-engines", "--immutable")
 
-    override fun getRemotePackageDetails(workingDir: File, packageName: String): PackageJson? {
+    internal open fun getRemotePackageDetails(workingDir: File, packageName: String): PackageJson? {
         yarnInfoCache.read(packageName)?.let { return parsePackageJson(it) }
 
         val process = run(workingDir, "info", "--json", packageName)
@@ -132,3 +433,15 @@ private fun extractDataNodes(output: String, type: String): Set<JsonElement> =
                 .mapNotNullTo(mutableSetOf()) { it["data"] }
         }
     }.getOrDefault(emptySet())
+
+private fun findDependencyModuleDir(dependencyName: String, searchModuleDirs: List<File>): List<File> {
+    searchModuleDirs.forEachIndexed { index, moduleDir ->
+        // Note: resolve() also works for scoped dependencies, e.g. dependencyName = "@x/y"
+        val dependencyModuleDir = moduleDir.resolve("node_modules/$dependencyName")
+        if (dependencyModuleDir.isDirectory) {
+            return listOf(dependencyModuleDir) + searchModuleDirs.subList(index, searchModuleDirs.size)
+        }
+    }
+
+    return emptyList()
+}
